fix validation for user data

+++ src/custom-hooks/useGetUserData.ts
@@ -11,10 +11,15 @@ export function useGetUserData() {
       try {
         const querySnapshot = await getDocs(collection(db, "users"));
         querySnapshot.forEach((e) => {
-          if (!e.data().backendTime || !e.data().name) {
-            deleteDoc(doc(db, "users", e.data().id));
-          } else {
-            temp.push(e.data());
+          if (
+            e.data().hasOwnProperty("backendTime") &&
+            e.data().hasOwnProperty("name")
+          ) {
+            if (!e.data().backendTime || !e.data().name) {
+              deleteDoc(doc(db, "users", e.data().id));
+            } else {
+              temp.push(e.data());
+            }
           }
         });
         setUserData([...temp]);
